cast

+++ src/ofApp.cpp
@@ -23,7 +23,7 @@ void ofApp::update()
 		room.setSSPin(0, false);
 		usleep(5000);
 
-		cout << "send :" << bt << endl;
+		cout << "send :" << (int)(bt) << endl;
 
 		room.setSSPin(0, true);
 		usleep(5000);
@@ -31,7 +31,7 @@ void ofApp::update()
 		spiSender::transfer(&bt);
 		room.setSSPin(0, false);
 
-		cout << "send :" << bt << endl;
+		cout << "send :" << (int)(bt) << endl;
 //		int i = 0;
 //
 //		room.sendSpi_single(0, 0x04, i);
@@ -60,14 +60,14 @@ void ofApp::update()
 		spiSender::transfer(&bt);
 		room.setSSPin(0, false);
 		usleep(5000);
-		cout << "send :" << bt << endl;
+		cout << "send :" << (int)(bt) << endl;
 
 		room.setSSPin(0, true);
 		usleep(5000);
 		bt = 0xFF;
 		spiSender::transfer(&bt);
 		room.setSSPin(0, false);
-		cout << "send :" << bt << endl;
+		cout << "send :" << (int)(bt) << endl;
 	}
 }
